+++ CHANGELOG.md
@@ -2,6 +2,23 @@
 
 Old Versions that are over 5 versions behind will be removed.
 
+## 1.2.0
+
+1. Maps
+   1. Added A Map class that is to be extended for code reusability
+   2. Made tilemap extend new Map Class
+2. Added ParticleContainer class
+3. Added new Effect class
+4. Preset Effects
+   1. Added Explosion preset effect
+   2. Added Smoke preset effect
+5. Scene
+   1. Added effect to add method
+   2. Added presetEffect object to add method
+6. Misc
+   1. Updated dependencies
+   2. Changed Version
+
 ## 1.1.0
 
 1. Game